feat: 1. 重构框架代码把偏业务侧代码转移到 agent-client-multiple-web 项目. 2. 业务侧代码增强智能体功能。3. 大量变动核心代码命名。
