+++ a2a-framework/a2a-agent-client/pom.xml
@@ -4,7 +4,7 @@
 	<parent>
 		<groupId>io.github.musaemotion</groupId>
 		<artifactId>a2a-framework</artifactId>
-		<version>0.2.0</version>
+		<version>0.2.1</version>
 		<relativePath>../pom.xml</relativePath>
 	</parent>
