+++ a2a-framework/a2a-common/src/main/java/com/musaemotion/a2a/common/AgentCard.java
@@ -39,8 +39,6 @@ import java.util.List;
 @NoArgsConstructor
 public class AgentCard implements Serializable {
 
-
-
     /**
      * 名称
      */
