+++ a2a-framework/a2a-host-agent/src/main/java/com/musaemotion/a2a/agent/host/core/SendTaskCallbackHandle.java
@@ -26,9 +26,8 @@ import com.musaemotion.a2a.common.base.Task;
  * @project：A2A
  * @date：2025/4/27 22:53
  * @description：HostAgent: fun(sendTask)-> obj(client: A2aRemoteAgentConnections).sendTask() 调用的回调
- * TODO 线程安全性？
  */
-public interface ISendTaskCallback {
+public interface SendTaskCallbackHandle {
 
 	/**
 	 * 发送任务回调，在前处理和后处理都有调用
