--- a2a-framework/a2a-host-agent/src/main/java/com/musaemotion/a2a/agent/host/manager/HostAgentManager.java
@@ -1,677 +1,0 @@
-/*
- * Copyright (c) 2025 MusaeMotion
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.musaemotion.a2a.agent.host.manager;
-
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.google.common.collect.Lists;
-import com.google.common.collect.Maps;
-import com.musaemotion.a2a.common.AgentCard;
-import com.musaemotion.a2a.common.event.TaskArtifactUpdateEvent;
-import com.musaemotion.a2a.common.event.TaskEvent;
-import com.musaemotion.a2a.common.event.TaskStatusUpdateEvent;
-import com.musaemotion.a2a.common.base.Common;
-import com.musaemotion.a2a.common.base.Task;
-import com.musaemotion.a2a.common.utils.JsonUtils;
-import com.musaemotion.a2a.agent.host.core.HostAgent;
-import com.musaemotion.a2a.agent.host.core.ISendTaskCallback;
-import com.musaemotion.a2a.agent.host.model.response.CommonMessageExt;
-import com.musaemotion.a2a.agent.host.model.response.PageInfo;
-import com.musaemotion.a2a.agent.host.model.response.PageUtils;
-import com.musaemotion.a2a.agent.host.model.response.SendMessageResponse;
-import com.musaemotion.a2a.agent.host.model.service.Conversation;
-import com.musaemotion.a2a.agent.host.model.service.RegisterAgentDto;
-import com.musaemotion.a2a.agent.host.model.service.SearchRemoteAgentDto;
-import com.musaemotion.a2a.agent.host.properties.A2aHostAgentProperties;
-import com.musaemotion.a2a.common.constant.MessageRole;
-import com.musaemotion.a2a.common.constant.TaskState;
-import com.musaemotion.a2a.common.utils.GuidUtils;
-import com.musaemotion.a2a.agent.client.server.PushNotificationServer;
-import com.musaemotion.agent.HostAgentPromptService;
-import com.musaemotion.agent.model.SendMessageRequest;
-import io.micrometer.observation.ObservationRegistry;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.ai.chat.messages.AssistantMessage;
-import org.springframework.ai.chat.model.ChatModel;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.event.EventListener;
-import org.springframework.data.domain.Page;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-import org.springframework.util.CollectionUtils;
-import org.springframework.util.StringUtils;
-import reactor.core.publisher.Flux;
-import reactor.core.publisher.FluxSink;
-
-import java.util.*;
-import java.util.concurrent.Callable;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
-import java.util.stream.Collectors;
-
-import static com.musaemotion.a2a.common.constant.ArtifactDataKey.*;
-import static com.musaemotion.a2a.common.constant.MetaDataKey.*;
-import static com.musaemotion.agent.BasisAgent.STATE;
-
-/**
- * @author：contact@musaemotion.com
- * @package：com.musaemotion.a2a.agent.host.service
- * @project：A2A
- * @date：2025/5/9 13:15
- * @description：请完善描述
- */
-@Slf4j
-@Service
-public class HostAgentManager implements ISendTaskCallback {
-
-	/**
-	 * 交谈管理器
-	 */
-	private AbstractConversationManager conversationManager;
-
-	/**
-	 * 消息管理器
-	 */
-	private AbstractMessageManager messageManager;
-
-	/**
-	 * 远程智能体管理器
-	 */
-	private AbstractRemoteAgentManager remoteAgentManager;
-
-	/**
-	 * 主机智能体
-	 */
-	private HostAgent hostAgent;
-
-	/**
-	 * 消息通知服务
-	 */
-	private PushNotificationServer pushNotificationServer;
-
-	/**
-	 * 智能体任务管理器
-	 */
-	private AbstractTaskCenterManager taskCenterManager;
-
-	/**
-	 * 可观察性
-	 */
-	private ObservationRegistry observationRegistry;
-
-	/**
-	 * host Agent 提示词service
-	 */
-	private HostAgentPromptService hostAgentPromptService;
-
-	/**
-	 * 接受通知
-	 * @param event
-	 */
-	@EventListener
-	public void handleNotification(String event) throws JsonProcessingException {
-		ObjectMapper mapper = new ObjectMapper();
-		Task task = mapper.readValue(event, Task.class);
-		log.info("收到消息: {}", event);
-		// 删除删除通知sse
-		SseEmitterManager.pushData(task.getSessionId(), task.getInputMessageId() , event);
-
-	}
-
-
-	/**
-	 *
-	 * @param chatModel
-	 * @param a2aHostAgentProperties
-	 * @param abstractConversationManager
-	 * @param abstractMessageManager
-	 * @param abstractRemoteAgentManager
-	 * @param pushNotificationServer
-	 */
-	@Autowired
-	public HostAgentManager(ChatModel chatModel, A2aHostAgentProperties a2aHostAgentProperties, AbstractConversationManager abstractConversationManager, AbstractMessageManager abstractMessageManager, AbstractRemoteAgentManager abstractRemoteAgentManager, AbstractTaskCenterManager abstractTaskCenterManager, HostAgentPromptService hostAgentPromptService, @Autowired(required = false)  PushNotificationServer pushNotificationServer,  @Autowired(required = false) ObservationRegistry observationRegistry) {
-		this.conversationManager = abstractConversationManager;
-		this.messageManager = abstractMessageManager;
-		this.remoteAgentManager = abstractRemoteAgentManager;
-		this.pushNotificationServer = pushNotificationServer;
-		this.taskCenterManager = abstractTaskCenterManager;
-		this.observationRegistry = observationRegistry;
-		this.hostAgentPromptService = hostAgentPromptService;
-		if (this.observationRegistry == null) {
-			this.observationRegistry = ObservationRegistry.NOOP;
-		}
-		this.hostAgent = HostAgent.builder()
-				.remoteAgentAddresses(
-						CollectionUtils.isEmpty(a2aHostAgentProperties.getRemoteAgentAddresses()) ?
-								Lists.newArrayList() : a2aHostAgentProperties.getRemoteAgentAddresses()
-				)
-				.pushNotificationServer(this.pushNotificationServer)
-				.remoteAgentManager(this.remoteAgentManager)
-				.taskCenterManager(this.taskCenterManager)
-				.messageManager(this.messageManager)
-				.observationRegistry(this.observationRegistry)
-				.hostAgentPromptService(this.hostAgentPromptService)
-				.chatModel(chatModel)
-				.sendTaskCallback(this)
-				.build();
-		this.hostAgent.initHostAgent();
-	}
-
-	/**
-	 * 检查参数
-	 * @param input
-	 */
-	private void checkSendMessageRequest(SendMessageRequest input) {
-		if (!StringUtils.hasText(input.getContent())) {
-			throw new IllegalArgumentException("content is null");
-		}
-		if (!StringUtils.hasText(input.getConversationId())) {
-			throw new IllegalArgumentException("conversationId is null");
-		}
-		if (!this.conversationManager.exist(input.getConversationId())) {
-			throw new IllegalArgumentException("conversation no exist for conversationId");
-		}
-	}
-
-
-	/**
-	 * 发送之前处理消息
-	 * @param input
-	 * @return
-	 */
-	private Common.Message sanitizeRequestMessage(SendMessageRequest input) {
-		Boolean isEmpty = Optional.ofNullable(input.getMetadata()).map(m -> m.isEmpty()).orElse(true);
-		Common.Message message = Common.Message.builder()
-				.role(MessageRole.USER)
-				.parts(input.getParams().getParts())
-				.metadata(isEmpty ? Maps.newConcurrentMap() : input.getMetadata())
-				.build();
-
-		Optional<Common.Message> optionalMessage = this.messageManager.lastByConversationId(input.getConversationId());
-		// 添加上一条消息
-		if (optionalMessage.isPresent()) {
-			message.getMetadata().put(LAST_MESSAGE_ID, optionalMessage.get().getMetadata().get(MESSAGE_ID));
-		}
-		return message;
-	}
-
-	/**
-	 * 处理响应消息
-	 * @param assistantMessage
-	 * @param userMessage
-	 * @return
-	 */
-	private Common.Message sanitizeResponseMessage(AssistantMessage assistantMessage, Common.Message userMessage, String messageId) {
-		Common.Message message = null;
-		if (JsonUtils.isJsonString(assistantMessage.getText())) {
-			message = Common.Message.builder()
-					.role(MessageRole.AGENT)
-					.metadata(assistantMessage.getMetadata())
-					.parts(Lists.newArrayList())
-					.build();
-			try {
-				JsonNode jsonNode = new ObjectMapper().readTree(assistantMessage.getText());
-				Common.Message finalMessage = message;
-				jsonNode.forEach(json -> {
-					if (json.has(ARTIFACT_FILE_URL) || json.has(ARTIFACT_FILE_ID)) {
-						Common.FilePart filePart = new Common.FilePart();
-						filePart.setFile(Common.FileContent.builder()
-								.uri(json.get(ARTIFACT_FILE_URL) == null ? "" : json.get(ARTIFACT_FILE_URL).textValue())
-								.name(json.get(ARTIFACT_FILE_ID) == null ? "" : json.get(ARTIFACT_FILE_ID).textValue())
-								.mimeType(json.get(ARTIFACT_MIME_TYPE) == null ? "" : json.get(ARTIFACT_MIME_TYPE).textValue())
-								.build());
-						finalMessage.getParts().add(filePart);
-						return;
-					}
-					if(json.has("type") && json.get("type").textValue().equals("text")) {
-						finalMessage.getParts().add(new ObjectMapper().convertValue(json, Common.TextPart.class));
-						return;
-					}
-					Common.DataPart dataPart = new Common.DataPart();
-					dataPart.setData(new ObjectMapper().convertValue(json, Map.class));
-					finalMessage.getParts().add(dataPart);
-
-				});
-			} catch (JsonProcessingException e) {
-				throw new RuntimeException("响应内容反序列化失败");
-			}
-		} else {
-			message = Common.Message.builder()
-					.role(MessageRole.AGENT)
-					.parts(Lists.newArrayList(new Common.TextPart(assistantMessage.getText())))
-					.metadata(assistantMessage.getMetadata())
-					.build();
-		}
-		message.getMetadata().put(LAST_MESSAGE_ID, userMessage.getMetadata().get(MESSAGE_ID).toString());
-		message.getMetadata().put(CONVERSATION_ID, userMessage.getMetadata().get(CONVERSATION_ID).toString());
-		message.getMetadata().put(MESSAGE_ID, StringUtils.hasText(messageId)? messageId: GuidUtils.createGuid());
-		return message;
-	}
-
-	/**
-	 * 构建工具上下文
-	 * @param input
-	 * @return
-	 */
-	private Map<String, Object> buildToolContext(SendMessageRequest input){
-		Map<String, Object> state = new HashMap<>();
-		state.put(INPUT_MESSAGE_METADATA, input.getMetadata());
-		state.put(SESSION_ID, input.getConversationId());
-		state.put(MAIN_TASK_ID, GuidUtils.createGuid());
-		Map<String, Object> toolContext = new HashMap<>();
-		toolContext.put(STATE, state);
-		return toolContext;
-	}
-
-	/**
-	 * 之前处理
-	 * @param input
-	 * @return
-	 */
-	private Common.Message sendBefore(SendMessageRequest input){
-		this.checkSendMessageRequest(input);
-		// 消息处理
-		Common.Message userMessage = this.sanitizeRequestMessage(input);
-		// 消息管理器保存
-		this.messageManager.upsert(userMessage);
-		return userMessage;
-	}
-
-
-	/**
-	 * 后处理
-	 * @param assistantMessage
-	 * @param userMessage
-	 * @return
-	 */
-	private Common.Message sendAfter(AssistantMessage assistantMessage, Common.Message userMessage){
-		return this.sendAfter(assistantMessage, userMessage, "");
-	}
-
-	/**
-	 * 后处理
-	 * @param assistantMessage
-	 * @param userMessage
-	 * @param messageId
-	 * @return
-	 */
-	private Common.Message sendAfter(AssistantMessage assistantMessage, Common.Message userMessage, String messageId)  {
-		Common.Message agnetMessage = this.sanitizeResponseMessage(assistantMessage, userMessage, messageId);
-		// 消息管理器保存
-		this.messageManager.upsert(agnetMessage);
-		return agnetMessage;
-	}
-
-	/**
-	 * 加载task信息
-	 * @param agnetMessage
-	 * @return
-	 */
-	private CommonMessageExt loadTask(Common.Message agnetMessage){
-		var message = CommonMessageExt.fromMessage(agnetMessage);
-		String lastMessageId = agnetMessage.getLastMessageId();
-		if(StringUtils.hasText(lastMessageId)){
-			List<Task> tasks = this.taskCenterManager.listByInputMessageId(Lists.newArrayList(agnetMessage.getLastMessageId()));
-			message.setTask(tasks);
-		}
-		return message;
-	}
-
-	/**
-	 * 同步请求消息
-	 * @param input
-	 * @return
-	 */
-	public SendMessageResponse<CommonMessageExt> call(SendMessageRequest input) {
-		Common.Message userMessage = this.sendBefore(input);
-		AssistantMessage assistantMessage = this.hostAgent.call(input, this.buildToolContext(input));
-		Common.Message agnetMessage = this.sendAfter(assistantMessage, userMessage);
-		var message = loadTask(agnetMessage);
-		// 删除删除通知sse
-		SseEmitterManager.removeEmitter(input.getConversationId(), input.getMessageId());
-		return SendMessageResponse.buildMessageResponse(
-				message,
-				input.getConversationId()
-		);
-	}
-
-	/**
-	 * 流请求包装消息
-	 * @param messages
-	 * @return
-	 */
-	private Common.Message streamFinishReasonMessage(List<Common.Message> messages){
-		var parts =	messages.stream().map(item->item.getParts()).flatMap(Collection::stream).collect(Collectors.toUnmodifiableList());
-		String text = parts.stream().filter(part->{ return  part instanceof Common.TextPart; }).map(item->((Common.TextPart) item).getText()).collect(Collectors.joining(""));
-		var agentMessage = messages.get(messages.size()-1);
-		agentMessage.getParts().clear();
-		agentMessage.getParts().add(new Common.TextPart(text));
-		return agentMessage;
-	}
-
-	/**
-	 * 发送流请求
-	 * @param input
-	 * @return
-	 */
-	public Flux<SendMessageResponse> stream(SendMessageRequest input) {
-		Common.Message userMessage = this.sendBefore(input);
-		Flux<AssistantMessage> fluxAssistantMessage = this.hostAgent.stream(input, this.buildToolContext(input));
-		String messageId = GuidUtils.createGuid();
-		List<Common.Message> messages = Lists.newArrayList();
-		return fluxAssistantMessage.doFinally(i -> {
-					// 删除删除通知sse
-					SseEmitterManager.removeEmitter(input.getConversationId(), input.getMessageId());
-					var agentMessage = this.streamFinishReasonMessage(messages);
-					this.messageManager.upsert(agentMessage);
-				})
-				.map(assistantMessage -> {
-					var agnetMessage = this.sanitizeResponseMessage(assistantMessage, userMessage, messageId);
-					try {
-						messages.add(agnetMessage);
-						// 加载任务消息
-						if ("STOP".equals(assistantMessage.getMetadata().get("finishReason"))) {
-							agnetMessage = loadTask(agnetMessage);
-						}
-
-					} catch (Exception e) {
-						agnetMessage.setParts(Lists.newArrayList(new Common.TextPart("智能体出现异常")));
-					}
-					return SendMessageResponse.buildMessageResponse(
-							agnetMessage,
-							input.getConversationId());
-				});
-	}
-
-	/**
-	 * 删除对话
-	 * @param conversationId
-	 */
-	@Transactional(rollbackFor = Exception.class)
-	public void deleteConversation(String conversationId) {
-		this.conversationManager.delete(conversationId);
-		this.messageManager.deleteByConversationId(conversationId);
-		this.taskCenterManager.deleteByConversationId(conversationId);
-	}
-
-	/**
-	 * 创建对话
-	 * @return
-	 */
-	public Conversation createConversation() {
-		return this.conversationManager.create("");
-	}
-
-	/**
-	 * 获取对话列表
-	 * @return
-	 */
-	public List<Conversation> listConversation() {
-		return this.conversationManager.list();
-	}
-
-	/**
-	 * 获取对话相关的消息
-	 * @return
-	 */
-	public List<CommonMessageExt> listMessage(String conversationId) {
-
-		var messages = this.messageManager.listByConversationId(conversationId);
-		List<CommonMessageExt> newMessages = messages.stream().map(item->CommonMessageExt.fromMessage(item)).collect(Collectors.toUnmodifiableList());
-
-		List<String> messageIds = messages.stream()
-				.filter(item->item.getLastMessageId()!=null)
-				.map(Common.Message::getLastMessageId).distinct()
-				.collect(Collectors.toUnmodifiableList());
-
-		List<Task> tasks = this.taskCenterManager.listByInputMessageId(messageIds);
-		newMessages.forEach(message->{
-			var ts =  tasks.stream().filter(task -> task.getInputMessageId().equals(message.getLastMessageId())).collect(Collectors.toUnmodifiableList());
-			if(!CollectionUtils.isEmpty(ts)){
-				message.setTask(ts);
-			}
-
-		});
-
-		return newMessages;
-	}
-
-	/**
-	 * 注册agent
-	 * @param input
-	 * @return
-	 */
-	public AgentCard registerAgent(RegisterAgentDto input) {
-		AgentCard agentCard = this.remoteAgentManager.registerAgent("http://" + input.getUrl());
-		if (this.pushNotificationServer != null) {
-			this.pushNotificationServer.registerAgent(agentCard.getName(), agentCard.getUrl());
-		}
-		return agentCard;
-	}
-
-	/**
-	 * Agent分页
-	 * @param searchInput
-	 * @param pageNum
-	 * @param pageSize
-	 * @return
-	 */
-	public PageInfo<AgentCard> pageListAgent(SearchRemoteAgentDto searchInput, Integer pageNum, Integer pageSize) {
-		pageNum = pageNum - 1;
-		Page<AgentCard> page = this.remoteAgentManager.pageList(searchInput, pageNum, pageSize);
-		PageInfo<AgentCard> pageInfo = PageUtils.springPageToMyPage(page);
-		return pageInfo;
-	}
-
-	/**
-	 * 删除agent
-	 * @param ids
-	 * @return
-	 */
-	public void deleteAgent(String ids) {
-		if (StringUtils.hasText(ids)) {
-			List<String> id = Arrays.stream(ids.split(",")).collect(Collectors.toList());
-			this.remoteAgentManager.delete(id);
-		}
-	}
-
-	/**
-	 * 根据对话id获取列表
-	 * @param conversationId
-	 */
-	public List<Task> listTaskByConversationId(String conversationId) {
-		return this.taskCenterManager.getByConversationId(conversationId);
-	}
-
-	/**
-	 * 删除 task
-	 * @param ids
-	 * @return
-	 */
-	public void deleteTask(String ids) {
-		if (StringUtils.hasText(ids)) {
-			List<String> id = Arrays.stream(ids.split(",")).collect(Collectors.toList());
-			this.taskCenterManager.removeTask(id);
-		}
-	}
-
-
-	/**
-	 *
-	 * @param newTask
-	 */
-	@Override
-	public void sendTaskCallback(Task newTask) {
-
-
-		// 添加消息栈关系，消息之间前后关系
-		this.insertIdTrace(newTask.getStatus().getMessage());
-
-		Optional<Task> optionalTask = this.taskCenterManager.getById(newTask.getId());
-		if (optionalTask.isPresent()) {
-			Task oldTask = optionalTask.get();
-			oldTask.setStatus(newTask.getStatus());
-			if(newTask.getArtifacts()!=null) {
-				// 返回多个产出结果工件
-				newTask.getArtifacts().forEach(artifact -> {
-					this.processArtifactEvent(oldTask, artifact);
-				});
-			}
-			this.insertMessageHistory(oldTask, newTask.getStatus().getMessage());
-			this.taskCenterManager.updateTask(oldTask);
-			return;
-		}
-		this.taskCenterManager.addTask(newTask);
-	}
-
-
-	/**
-	 * 更新任务状态
-	 * @param taskStatusUpdateEvent
-	 */
-	@Override
-	public void sendTaskCallback(TaskStatusUpdateEvent taskStatusUpdateEvent) {
-
-		Task oleTask = this.addOrGetTask(taskStatusUpdateEvent);
-
-		oleTask.setStatus(taskStatusUpdateEvent.getStatus());
-
-		// 添加消息栈关系，消息之间前后关系
-		this.insertIdTrace(taskStatusUpdateEvent.getStatus().getMessage());
-		// 给当前任务添加历史记录
-		this.insertMessageHistory(oleTask, taskStatusUpdateEvent.getStatus().getMessage());
-
-		this.taskCenterManager.updateTask(oleTask);
-	}
-
-	/**
-	 * 完成任务状态更新
-	 * @param taskArtifactUpdateEvent
-	 */
-	@Override
-	public void sendTaskCallback(TaskArtifactUpdateEvent taskArtifactUpdateEvent) {
-
-		Task oldTask = this.addOrGetTask(taskArtifactUpdateEvent);
-
-		this.processArtifactEvent(oldTask, taskArtifactUpdateEvent.getArtifact());
-
-		this.taskCenterManager.updateTask(oldTask);
-	}
-
-	/**
-	 * 消息栈
-	 * @param message
-	 */
-	private void insertIdTrace(Common.Message message) {
-		if (message == null) {
-			return;
-		}
-		String messageId = message.getMessageId();
-		String lastMessageId = message.getLastMessageId();
-		if (messageId != null && lastMessageId != null) {
-			// 建立上一条和下一条的关系
-			this.taskCenterManager.getNextId().put(lastMessageId, messageId);
-		}
-	}
-
-	/**
-	 * 状态更新，有可能是初始化创建
-	 * @param input
-	 * @return
-	 */
-	private Task addOrGetTask(TaskEvent input) {
-		Optional<Task> optionalTask = this.taskCenterManager.getById(input.getId());
-		if (optionalTask.isEmpty()) {
-			String conversationId = null;
-			if (input.getMetadata() != null && input.getMetadata().containsKey(CONVERSATION_ID)) {
-				conversationId = input.getMetadata().get(CONVERSATION_ID).toString();
-			}
-			Task newTask = Task.builder().id(input.getId())
-					.status(Common.TaskStatus.builder().state(TaskState.SUBMITTED).build())
-					.metadata(input.getMetadata())
-					.artifacts(Lists.newArrayList())
-					.sessionId(conversationId)
-					.build();
-			this.taskCenterManager.addTask(newTask);
-			return newTask;
-		}
-
-		return optionalTask.get();
-	}
-
-	/**
-	 * 插入聊天记录
-	 * @param oldTask
-	 * @param message
-	 */
-	private void insertMessageHistory(Task oldTask, Common.Message message) {
-		if (message == null || message.getMessageId() == null) {
-			return;
-		}
-		if (oldTask.getHistory() == null) {
-			oldTask.setHistory(new ArrayList<>());
-		}
-		// 状态消息的消息id
-		String newMessageId = message.getMessageId();
-		if (!oldTask.getHistory().stream().anyMatch(historicalMessage -> historicalMessage.getMessageId().equals(newMessageId))) {
-			// TaskState.COMPLETED 角色为空，是消息完成消息
-			if(message.getRole() == null) {
-				message = Common.Message.newMessage(MessageRole.AGENT, Lists.newArrayList(new Common.TextPart("完成请求更新状态")), Maps.newHashMap());
-			}
-			oldTask.getHistory().add(message);
-		} else {
-			log.warn("Message id already in history: {} , History: {}", newMessageId, oldTask.getHistory());
-		}
-	}
-
-	/**
-	 * 处理工件, 后续调整到任务中心去取对应工件信息
-	 * @param currentTask
-	 * @param artifact
-	 */
-	private void processArtifactEvent(Task currentTask, Common.Artifact artifact) {
-		if (!artifact.getAppend()) {
-			if (artifact.getLastChunk() || artifact.getLastChunk() == null) {
-				if (currentTask.getArtifacts() == null) {
-					currentTask.setArtifacts(new ArrayList<>());
-				}
-				currentTask.getArtifacts().add(artifact);
-			} else {
-				if (!this.taskCenterManager.getArtifactChunks().containsKey(currentTask.getId())) {
-					Map<Integer, Common.Artifact> artifactMap = Maps.newHashMap();
-					this.taskCenterManager.getArtifactChunks().put(
-							currentTask.getId(),
-							artifactMap
-					);
-				}
-				Map<Integer, Common.Artifact> artifactMap = this.taskCenterManager.getArtifactChunks().get(currentTask.getId());
-				artifactMap.put(artifact.getIndex(), artifact);
-			}
-		} else {
-			Map<Integer,  Common.Artifact> tempArtifacts =  this.taskCenterManager.getArtifactChunks().get(currentTask.getId());
-			if (tempArtifacts != null && tempArtifacts.containsKey(artifact.getIndex())) {
-				Common.Artifact currentTempArtifact = tempArtifacts.get(artifact.getIndex());
-				currentTempArtifact.getParts().addAll(artifact.getParts());
-				if (artifact.getLastChunk()) {
-					currentTask.getArtifacts().add(currentTempArtifact);
-					tempArtifacts.remove(artifact.getIndex());
-				}
-			}
-		}
-	}
-}
