--- a2a-framework/a2a-host-agent/src/main/java/com/musaemotion/a2a/agent/host/model/service/AgentCardVo.java
@@ -1,34 +1,0 @@
-/*
- * Copyright (c) 2025 MusaeMotion
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.musaemotion.a2a.agent.host.model.service;
-
-import com.musaemotion.a2a.common.AgentCard;
-import lombok.Data;
-import lombok.EqualsAndHashCode;
-
-/**
- * @author：contact@musaemotion.com
- * @package：com.a2a.demo.agent.client.model
- * @project：A2A
- * @date：2025/5/6 15:48
- * @description：请完善描述
- */
-@Data
-@EqualsAndHashCode(callSuper = true)
-public class AgentCardVo extends AgentCard {
-    private String id;
-}
