+++ a2a-framework/a2a-host-agent/src/main/java/com/musaemotion/a2a/agent/host/model/service/RegisterAgentDto.java
@@ -27,5 +27,6 @@ import lombok.Data;
  */
 @Data
 public class RegisterAgentDto {
+	// 注入 url
     private String url;
 }
