+++ a2a-framework/pom.xml
@@ -3,7 +3,7 @@
 
     <modelVersion>4.0.0</modelVersion>
 	<groupId>io.github.musaemotion</groupId>
-	<version>0.2.0</version>
+	<version>0.2.1</version>
     <artifactId>a2a-framework</artifactId>
     <packaging>pom</packaging>
