+++ agent-client-multiple-web/README.md
@@ -10,7 +10,7 @@ If you want to implement your own host-agent, simply add the following Maven dep
 <dependency>
     <groupId>io.github.musaemotion</groupId>
     <artifactId>a2a-host-agent</artifactId>
-    <version>0.2.0</version>
+    <version>0.2.1</version>
 </dependency>
 ```
