+++ agent-client-multiple-web/README.zh-CN.md
@@ -9,7 +9,7 @@
     <dependency>
         <groupId>io.github.musaemotion</groupId>
         <artifactId>a2a-host-agent</artifactId>
-        <version>0.2.0</version>
+        <version>0.2.1</version>
     </dependency>
 ```
