+++ agent-client-multiple-web/pom.xml
@@ -4,7 +4,7 @@
     <parent>
         <groupId>io.github.musaemotion</groupId>
 		<artifactId>a2a</artifactId>
-		<version>0.2.0</version>
+		<version>0.2.1</version>
 		<relativePath>../pom.xml</relativePath>
     </parent>
     <groupId>com.musaemotion.demo</groupId>
@@ -56,7 +56,7 @@
             <artifactId>spring-ai-starter-model-ollama</artifactId>
         </dependency>
 
-        <dependency>
+		<dependency>
             <groupId>com.vladmihalcea</groupId>
             <artifactId>hibernate-types-60</artifactId>
         </dependency>
