+++ agent-client-multiple-web/src/main/java/com/a2a/demo/agent/client/service/MysqlConversationManager.java
@@ -18,7 +18,7 @@ package com.a2a.demo.agent.client.service;
 
 import com.a2a.demo.agent.client.entity.ConversationEntity;
 import com.a2a.demo.agent.client.repository.ConversationRepository;
-import com.musaemotion.a2a.agent.host.model.service.Conversation;
+import com.a2a.demo.agent.client.dto.Conversation;
 import com.musaemotion.a2a.agent.host.manager.AbstractConversationManager;
 import jakarta.transaction.Transactional;
 import lombok.extern.slf4j.Slf4j;
@@ -54,8 +54,11 @@ public class MysqlConversationManager extends AbstractConversationManager {
     }
 
 
-
-    @Override
+	/**
+	 * 创建交谈
+	 * @param conversationName
+	 * @return
+	 */
     public Conversation create(String conversationName) {
         ConversationEntity conversationEntity = ConversationEntity.newConversation(conversationName);
         this.repository.save(conversationEntity);
@@ -72,11 +75,10 @@ public class MysqlConversationManager extends AbstractConversationManager {
         return this.repository.existsById(conversationId);
     }
 
-    /**
-     * 删除交谈对话
-     * @param conversationId
-     */
-    @Override
+	/**
+	 * 删除一个交谈信息
+	 * @param conversationId
+	 */
     @Transactional(rollbackOn = Exception.class)
     public void delete(String conversationId) {
        this.repository.deleteById(conversationId);
@@ -86,7 +88,6 @@ public class MysqlConversationManager extends AbstractConversationManager {
      * 获取交谈列表
      * @return
      */
-    @Override
     public List<Conversation> list() {
         return this.repository.findAll().stream().map(item -> item.toDto()).collect(Collectors.toList());
     }
