+++ agent-client-multiple-web/src/main/resources/db/migration/V1.0.2__agent_alter.sql
@@ -1,0 +1,2 @@
+ALTER TABLE `a2a`.`remote_agent`
+    ADD COLUMN `enable` TINYINT(1) NULL AFTER `agent_card`;
