+++ agent-server-ollama-sentiment-analysis/README.zh-CN.md
@@ -8,7 +8,7 @@
 <dependency>
     <groupId>io.github.musaemotion</groupId>
     <artifactId>a2a-agent-server</artifactId>
-    <version>0.2.0</version>
+    <version>0.2.1</version>
 </dependency>
 ```
