+++ agent-server-ollama-sentiment-analysis/pom.xml
@@ -5,7 +5,7 @@
     <parent>
 		<groupId>io.github.musaemotion</groupId>
 		<artifactId>a2a</artifactId>
-		<version>0.2.0</version>
+		<version>0.2.1</version>
         <relativePath>../pom.xml</relativePath>
     </parent>
     <groupId>com.musaemotion.demo</groupId>
@@ -31,12 +31,12 @@
             <version>${project.version}</version>
         </dependency>
 
-        <dependency>
-            <groupId>org.springframework.ai</groupId>
-            <artifactId>spring-ai-starter-model-ollama</artifactId>
-        </dependency>
+		<dependency>
+			  <groupId>org.springframework.ai</groupId>
+			  <artifactId>spring-ai-starter-model-ollama</artifactId>
+		</dependency>
 
-    </dependencies>
+	</dependencies>
 
     <build>
 		<finalName>agent-sentiment-analysis</finalName>
