+++ agent-server-zhipu-generate-images/pom.xml
@@ -4,7 +4,7 @@
     <parent>
 		<groupId>io.github.musaemotion</groupId>
 		<artifactId>a2a</artifactId>
-		<version>0.2.0</version>
+		<version>0.2.1</version>
     </parent>
 
     <artifactId>agent-server-zhipu-generate-images</artifactId>
