+++ pom.xml
@@ -36,7 +36,7 @@
     </properties>
     <groupId>io.github.musaemotion</groupId>
     <artifactId>a2a</artifactId>
-    <version>0.2.0</version>
+    <version>0.2.1</version>
     <name>A2A</name>
     <description>A2A</description>
     <dependencyManagement>
